+++ CHANGELOG.md
@@ -6,6 +6,13 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
+* New components:
+  * `<cc-addon-redis-options>`
+  * `<cc-addon-jenkins-options>`
+
+* `<cc-addon-elasticsearch-options>`
+  * Add encryption at rest option
+
 * `cc-addon-encryption-at-rest-option` template
   * Drop documentation URL parameter